+++ src/config/networks.ts
@@ -10,24 +10,8 @@ import WestendInlineSVG from 'img/westend_inline.svg?react';
 import PolkadotTokenSVG from 'config/tokens/svg/DOT.svg?react';
 import KusamaTokenSVG from 'config/tokens/svg/KSM.svg?react';
 import WestendTokenSVG from 'config/tokens/svg/WND.svg?react';
-
-import type { NetworkName, Networks } from 'types';
+import type { Networks, SystemChain } from 'types';
 import BigNumber from 'bignumber.js';
-
-// DEPRECATION: Paged Rewards
-//
-// Temporary until paged rewards migration has completed on all networks. Wait 84 eras from Polkadot
-// start: 1420 + 84 = 1504, when full history depth will be moved over to new paged rewards storage.
-export const NetworksWithPagedRewards: NetworkName[] = [
-  'polkadot',
-  'kusama',
-  'westend',
-];
-export const PagedRewardsStartEra: Record<NetworkName, BigNumber | null> = {
-  polkadot: new BigNumber(1420),
-  kusama: new BigNumber(6514),
-  westend: new BigNumber(7167),
-};
 
 export const NetworkList: Networks = {
   polkadot: {
@@ -46,7 +30,6 @@ export const NetworkList: Networks = {
         Stakeworld: 'wss://dot-rpc.stakeworld.io',
       },
     },
-    namespace: '91b171bb158e2d3848fa23a9f1c25182',
     colors: {
       primary: {
         light: 'rgb(211, 48, 121)',
@@ -103,7 +86,6 @@ export const NetworkList: Networks = {
         Stakeworld: 'wss://ksm-rpc.stakeworld.io',
       },
     },
-    namespace: 'b0a8d493285c2df73290dfb7e61f870f',
     colors: {
       primary: {
         light: 'rgb(31, 41, 55)',
@@ -159,7 +141,6 @@ export const NetworkList: Networks = {
         Stakeworld: 'wss://wnd-rpc.stakeworld.io',
       },
     },
-    namespace: 'e143f23803ac50e8f6f8e62695d1ce9e',
     colors: {
       primary: {
         light: '#da4e71',
@@ -199,5 +180,44 @@ export const NetworkList: Networks = {
     },
     defaultFeeReserve: 0.1,
     maxExposurePageSize: new BigNumber(64),
+  },
+};
+
+export const SystemChainList: Record<string, SystemChain> = {
+  'people-polkadot': {
+    name: 'people-polkadot',
+    ss58: 0,
+    units: 10,
+    unit: 'DOT',
+    endpoints: {
+      lightClient: 'people_polkadot', // NOTE: Currently not being used. TODO: Revise this and activate once People chain specs are available to use.
+      rpcEndpoints: {
+        Parity: 'wss://polkadot-people-rpc.polkadot.io',
+      },
+    },
+  },
+  'people-kusama': {
+    name: 'people-kusama',
+    ss58: 2,
+    units: 12,
+    unit: 'KSM',
+    endpoints: {
+      lightClient: 'people_kusama', // NOTE: Currently not being used. TODO: Revise this and activate once People chain specs are available to use.
+      rpcEndpoints: {
+        Parity: 'wss://kusama-people-rpc.polkadot.io',
+      },
+    },
+  },
+  'people-westend': {
+    name: 'people-westend',
+    ss58: 42,
+    units: 12,
+    unit: 'WND',
+    endpoints: {
+      lightClient: 'people_westend', // NOTE: Currently not being used. TODO: Revise this and activate once People chain specs are available to use.
+      rpcEndpoints: {
+        Parity: 'wss://westend-people-rpc.polkadot.io',
+      },
+    },
   },
 };