+++ src/controllers/Subscriptions/types.ts
@@ -1,0 +1,17 @@
+// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+import type { AnyJson } from '@w3ux/types';
+
+// Define all possible subscription classes.
+// TODO: Add subscription classes here.
+export type Subscription = AnyJson;
+
+// the record of subscriptions, keyed by tabId.
+export type ChainSubscriptions = Record<string, Subscription>;
+
+// Abstract class that ensures all subscription classes have an unsubscribe method.
+export abstract class Unsubscribable {
+  // Unsubscribe from unsubs present in this class.
+  abstract unsubscribe: () => void;
+}