+++ src/model/Api/types.ts
@@ -1,8 +1,7 @@
 // Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type * as ScType from '@substrate/connect';
-import type { NetworkName } from 'types';
+import type { NetworkName, SystemChainId } from 'types';
 
 export interface APIConfig {
   type: ConnectionType;
@@ -11,15 +10,11 @@ export interface APIConfig {
 
 export interface APIEventDetail {
   status: EventApiStatus;
-  network: NetworkName;
-  type: ConnectionType;
+  network: NetworkName | SystemChainId;
+  chainType: ApiChainType;
+  connectionType: ConnectionType;
   rpcEndpoint: string;
   err?: string;
-}
-
-export interface SubstrateConnect {
-  WellKnownChain: (typeof ScType)['WellKnownChain'];
-  createScClient: (typeof ScType)['createScClient'];
 }
 
 export type ConnectionType = 'ws' | 'sc';
@@ -27,3 +22,5 @@ export type ConnectionType = 'ws' | 'sc';
 export type ApiStatus = 'connecting' | 'connected' | 'disconnected' | 'ready';
 
 export type EventApiStatus = ApiStatus | 'error';
+
+export type ApiChainType = 'relay' | 'system';