+++ vite.config.ts
@@ -23,13 +23,6 @@ export default defineConfig({
   ],
   build: {
     outDir: 'build',
-    rollupOptions: {
-      output: {
-        manualChunks: {
-          '@substrate/connect': ['@substrate/connect'],
-        },
-      },
-    },
   },
   server: {
     fs: {