+++ docs/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [1.6.0](https://github.com/polkadot-cloud/polkadot-staking-dashboard/compare/v1.5.0...v1.6.0) (2024-08-01)
+
+
+### Features
+
+* Activate People Chain and re-enable identities ([#2198](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2198)) ([2a6a624](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/2a6a6245037a52e4a2666712d77c48cc7eabd3c6))
+* Check if Polkadot app is at least generic version ([#2191](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2191)) ([4aea706](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/4aea706af315d816c0c31003b46b5139937b0b98))
+* Generic Ledger app support ([#2181](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2181)) ([e4b2c0c](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/e4b2c0cc13490605a148578f4f79cbd809042d66))
+* Misc styling ([#2157](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2157)) ([e5f2537](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/e5f2537e35b791565f4e652db622ce48b0722a33))
+* New ledger app tx support ([#2165](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2165)) ([71b218f](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/71b218fc6a8f087f31e66a4499d540cf03d2a768))
+* Nominee decentralization: How decentralized is your nomination? ([#2185](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2185)) ([7dd170a](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/7dd170a7a50cf1e6b41c5f68d65fcf9743acc8da))
+* **refactor:** `@substrate/connect` updates ([#2196](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2196)) ([0a868b6](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/0a868b6c16cc0d0485ea8771c18e70a84ec1639a))
+* **refactor:** Abstract state bootstrapping & subscriptions from Api ([#2193](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2193)) ([8accc8b](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/8accc8bc5ba0ac8c39bfd0b5815ad85fd3b08991))
+* **refactor:** Controller renames ([#2194](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2194)) ([e2a6829](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/e2a682948ba3e19032b753135b8dc684d1eaf84e))
+* **refactor:** disable identities on networks where People Chain is being used ([#2183](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2183)) ([e5edec2](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/e5edec27f20c3a26d1c3398195311ccd49cfd57e))
+* **refactor:** People chain on Polkadot - disable identities on Relay Chain ([#2192](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2192)) ([6097a12](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/6097a12b36177fbf84f1cf565233d3f36b72a6d9))
+* **refactor:** Remove pre-paged rewards deprecation logic ([#2197](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2197)) ([c4fada1](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/c4fada1e5a616b48fe5c3327c14282eff8f13d75))
+* **refactor:** Use `merkelise-metadata` for Ledger signing, discontinue metadata service ([#2195](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2195)) ([907ea60](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/907ea603d062af4b0fdaee762e69d106da08b0ac))
+* update favicon ([#2162](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2162)) ([830b126](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/830b1266b2becbd2d5782c3721b6252b8a9803ba))
+* update header logo ([#2161](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2161)) ([e640d21](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/e640d210003afcd67760e7ab63457e3aff3bbfa9))
+
+
+### Bug Fixes
+
+* ledger hardware patch ([#2186](https://github.com/polkadot-cloud/polkadot-staking-dashboard/issues/2186)) ([038219f](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/038219f4305c48aca3bf6209b0727e05da98d76c))
+* vault payload ([b431d78](https://github.com/polkadot-cloud/polkadot-staking-dashboard/commit/b431d78c07c12723a976be6f997d061d505321f8))
+
+
 ## [1.5.0](https://github.com/polkadot-cloud/polkadot-staking-dashboard/compare/v1.4.4...v1.5.0) (2024-06-12)